lzseq: distance-dependent MIN_MATCH and wider MatchCandidate (Phase 4)
Reject short matches at large distances where code+extra-bits cost
exceeds literal cost. Add min_profitable_length(offset) to lzseq with
tiered thresholds (close=3, medium=4, far=5+). Integrate into lazy
matching loop.

Widen MatchCandidate to u32 offset/length for LzSeq support. Add
distance-aware match_cost() to optimal CostModel using actual
code+extra-bits costs instead of fixed overhead.

+++ src/webgpu/lz77.rs
@@ -933,8 +933,8 @@ impl WebGpuEngine {
             for k in 0..TOPK_K {
                 let p = packed[pos * TOPK_K + k];
                 slot[k] = MatchCandidate {
-                    offset: (p & 0xFFFF) as u16,
-                    length: (p >> 16) as u16,
+                    offset: p & 0xFFFF,
+                    length: p >> 16,
                 };
             }
         }
